requirements.txt
torch
torchvision
matplotlib
seaborn
scikit-learn
pillow
numpy
